Move verify and sign request to signature.
