+++ lib/client.js
@@ -6,11 +6,6 @@
 const _ = require('lodash');
 const {httpsAgent} = require('bedrock-https-agent');
 const {httpClient} = require('@digitalbazaar/http-client');
-const {
-  createAuthzHeader,
-  createSignatureString
-} = require('http-signature-header');
-const {createHeaderValue} = require('@digitalbazaar/http-digest-header');
 const https = require('https');
 const bedrock = require('bedrock');
 const {config, util: {BedrockError}} = bedrock;
@@ -19,7 +14,7 @@ const {pipeline} = require('stream');
 const split2 = require('split2');
 const {validate} = require('bedrock-validation');
 const {URL} = require('url');
-const {getKeyPair} = require('./signature');
+const {signRequest, getKeyPair} = require('./signature');
 
 let EVENTS_VALIDATION_HTTPS_AGENT;
 const EVENTS_VALIDATION_URL_CACHE = new LRU({max: 100, maxAge: 60 * 1000});
@@ -135,7 +130,7 @@ exports.notifyPeer = async ({ledgerNodeId, remotePeer}) => {
   // FIXME: send actual gossip URL instead of `localPeerId` twice
   const json = {peer: {id: key.id, url: key.id}};
   try {
-    const headers = await _signRequest({url, json, signer: key.signer()});
+    const headers = await signRequest({url, json, signer: key.signer()});
     await httpClient.post(url, {
       agent: httpsAgent,
       headers,
@@ -274,47 +269,4 @@ function _processHTTPError(error) {
     cause = new BedrockError(error.message, 'NetworkError');
   }
   return {cause, httpStatusCode};
-}
-
-async function _signRequest({url, json, signer}) {
-  const created = Math.floor(Date.now() / 1000);
-  const expires = created + 5 * 60;
-  const includeHeaders = [
-    '(created)',
-    '(expires)',
-    // (key-id) will include the peerId in the signature itself
-    '(key-id)',
-    '(request-target)',
-    'digest',
-    'content-type',
-    'host'
-  ];
-  const headers = {
-    digest: await createHeaderValue({data: json, useMultihash: true}),
-    'content-type': 'application/json'
-  };
-  const plainText = createSignatureString({
-    includeHeaders,
-    headers,
-    requestOptions: {
-      created,
-      expires,
-      url,
-      method: 'POST',
-      headers,
-      keyId: signer.id
-    }
-  });
-  const data = new TextEncoder().encode(plainText);
-  const sigBuffer = await signer.sign({data});
-  const signature = Buffer.from(
-    sigBuffer, sigBuffer.offset, sigBuffer.length).toString('base64');
-  const authorization = createAuthzHeader({
-    created,
-    expires,
-    includeHeaders,
-    keyId: signer.id,
-    signature
-  });
-  return {authorization, ...headers};
 }
