+++ lib/server.js
@@ -5,6 +5,7 @@
 
 const _events = require('./events');
 const _peers = require('./peers');
+const _signature = require('./signature');
 const _peerEvents = require('./peerEvents');
 const _history = require('./history');
 const _localPeers = require('./localPeers');
@@ -12,11 +13,8 @@ const {asyncHandler} = require('bedrock-express');
 const bedrock = require('bedrock');
 const bodyParser = require('body-parser');
 const brLedgerNode = require('bedrock-ledger-node');
-const {parseRequest} = require('http-signature-header');
-const {Ed25519VerificationKey2018} =
-  require('@digitalbazaar/ed25519-verification-key-2018');
 const brRest = require('bedrock-rest');
-const {config, util: {BedrockError}} = bedrock;
+const {config} = bedrock;
 const {callbackify} = require('util');
 const logger = require('./logger');
 const {validate} = require('bedrock-validation');
@@ -120,7 +118,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     routes.notify, brRest.when.prefers.ld,
     validate('continuity-server.notification'),
     asyncHandler(async (req, res) => {
-      const {verified, keyId} = await _verifyRequest({req});
+      const {verified, keyId} = await _signature.verifyRequest({req});
       if(!(verified && req.body.peer.id === keyId)) {
         return res.status(403).end();
       }
@@ -130,44 +128,6 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       res.status(204).end();
     }));
 });
-
-async function _verifyRequest({req}) {
-  try {
-    const {url, method, headers} = req;
-    const expectedHeaders = [
-      '(created)',
-      '(expires)',
-      // (key-id) will include the peerId in the signature itself
-      '(key-id)',
-      '(request-target)',
-      'digest',
-      'content-type',
-      'host'
-    ];
-    const parsed = parseRequest(
-      {url, method, headers}, {headers: expectedHeaders});
-    // get parsed parameters from from HTTP header and generate signing string
-    const {keyId, signingString, params: {signature: b64Signature}} = parsed;
-    // returns the key bytes
-    const key = new Ed25519VerificationKey2018({
-      id: keyId,
-      type: 'Ed25519VerificationKey2018',
-      controller: keyId,
-      publicKeyBase58: await _localPeers.getPublicKeyFromId({peerId: keyId})
-    });
-    const verifier = key.verifier();
-    const encoder = new TextEncoder();
-    const data = encoder.encode(signingString);
-    const signature = Buffer.from(b64Signature, 'base64');
-    // verify HTTP signature
-    const verified = await verifier.verify({data, signature});
-    return {verified, keyId};
-  } catch(e) {
-    throw new BedrockError(
-      'Signature verification failed.', 'DataError',
-      {public: true, httpStatusCode: 403}, e);
-  }
-}
 
 async function _handleNotification({req}) {
   const {peer} = req.body;
