+++ lib/signature.js
@@ -8,6 +8,12 @@ const _localPeers = require('./localPeers');
 const bedrock = require('bedrock');
 const {config: {constants}, util: {BedrockError}} = bedrock;
 const jsigs = require('jsonld-signatures');
+const {
+  parseRequest,
+  createAuthzHeader,
+  createSignatureString
+} = require('http-signature-header');
+const {createHeaderValue} = require('@digitalbazaar/http-digest-header');
 const {documentLoader} = require('bedrock-jsonld-document-loader');
 const logger = require('./logger');
 const {Ed25519VerificationKey2018} =
@@ -53,6 +59,67 @@ exports.sign = async ({event, ledgerNodeId}) => {
   return {signed, peerId: key.id};
 };
 
+const coveredContent = [
+  '(created)',
+  '(expires)',
+  // (key-id) will include the peerId in the signature itself
+  '(key-id)',
+  '(request-target)',
+  'digest',
+  'content-type',
+  'host'
+];
+
+/**
+ * Signs a request with a valid HTTP Signature header.
+ *
+ * @param {object} options - Options to use.
+ * @param {URL} options.url - The request URL.
+ * @param {object} options.json - The request body.
+ * @param {object} options.signer - A signer for the signature.
+ * @param {Array<string>} [options.includeHeaders = coveredContent] - Which
+ *   headers to include in the signature.
+ *
+ * @returns {Promise<object>} Returns the headers for the request.
+ */
+exports.signRequest = async ({
+  url,
+  json,
+  signer,
+  includeHeaders = coveredContent
+}) => {
+  const created = Math.floor(Date.now() / 1000);
+  const expires = created + 5 * 60;
+  const headers = {
+    digest: await createHeaderValue({data: json, useMultihash: true}),
+    'content-type': 'application/json'
+  };
+  const plainText = createSignatureString({
+    includeHeaders,
+    headers,
+    requestOptions: {
+      created,
+      expires,
+      url,
+      method: 'POST',
+      headers,
+      keyId: signer.id
+    }
+  });
+  const data = new TextEncoder().encode(plainText);
+  const sigBuffer = await signer.sign({data});
+  const signature = Buffer.from(
+    sigBuffer, sigBuffer.offset, sigBuffer.length).toString('base64');
+  const authorization = createAuthzHeader({
+    created,
+    expires,
+    includeHeaders,
+    keyId: signer.id,
+    signature
+  });
+  return {authorization, ...headers};
+};
+
 /**
  * Verify the proof signature on a merge event using the public key derived
  * from its peer ID. This method assumes that the merge event has a single
@@ -95,6 +162,44 @@ exports.verify = async ({event}) => {
   }
 
   return {controller: publicKeyId};
+};
+
+/**
+ * Verifies a request's HTTP signature header.
+ *
+ * @param {object} options to use.
+ * @param {object} options.req - An express Request object.
+ * @param {Array<string>} [options.expectedHeaders = coveredContent] - Which
+ *   headers are expected to be in the request.
+ *
+ * @returns {Promise<object>} The result of verification and the keyId.
+ */
+exports.verifyRequest = async ({req, expectedHeaders = coveredContent}) => {
+  try {
+    const {url, method, headers} = req;
+    const parsed = parseRequest(
+      {url, method, headers}, {headers: expectedHeaders});
+    // get parsed parameters from from HTTP header and generate signing string
+    const {keyId, signingString, params: {signature: b64Signature}} = parsed;
+    // returns the key bytes
+    const key = new Ed25519VerificationKey2018({
+      id: keyId,
+      type: 'Ed25519VerificationKey2018',
+      controller: keyId,
+      publicKeyBase58: await _localPeers.getPublicKeyFromId({peerId: keyId})
+    });
+    const verifier = key.verifier();
+    const encoder = new TextEncoder();
+    const data = encoder.encode(signingString);
+    const signature = Buffer.from(b64Signature, 'base64');
+    // verify HTTP signature
+    const verified = await verifier.verify({data, signature});
+    return {verified, keyId};
+  } catch(e) {
+    throw new BedrockError(
+      'Signature verification failed.', 'DataError',
+      {public: true, httpStatusCode: 403}, e);
+  }
 };
 
 function _getPublicKey(publicKeyId) {
